content: update instruction on Getting Started
Previous instruction assume that user will clone the repository and start
working from it.  This will cause many import error due to user may
clone on different repository name.

The new instruction require user to setup new repository and import
the ciigo package to use it.  This method is much flexible, give an
option to update the package as user wish.