all: change the Convert function to use type ConvertOptions
Previously, we pass the directory to be scanned for asciidoc markup files
and path to HTML template on Convert function.  Adding new option to
Convert will cause changes on the Convert signature.

To prevent this, we changes the Convert signature from multiple parameters
into single parameter ConvertOptions.

While at it, change the variable name HTMLTemplate to HtmlTemplate.