all: change the Serve signature to ServeOptions
Previously, we pass four parameters to Serve function: the instance
to memfs.MemFS, the root directory, the address to listen, and
path to HTML template.

In case we need to add new parameter in the future, the Serve function
signature will changes and this is not good for consumer of API.

This commit changes the Serve function parameters to ServeOptions
so we can add optional parameter in the future without changes to its
signature.