all: check for excluded file before processing sub directory
Previously, if the file path match with one of the excluded pattern,
we keep processing the sub directory to find the markup files.
This may cause an error "too many open files" if excluded directory
contains many sub directory and/or files.

This changes fix this issue by checking the path with excluded pattern
first before diving into sub directory.