all: try to fix test that sometimes fail inside container
When the test running we create directory testdata/watcher, create
a new file testdata/watcher/index.adoc and expect that the modify time
for testdata/watcher is changes.

Except that sometimes it is not. The modification time of directory
watcher before and after the file created most of times equal and this
cause the test wait indifinitely and fail.

This changes add 1 second delay before creating file inside directory
to make sure that the modification time changes.