go.mod: update module asciidoctor-go and share to latest release
The asciidoctor-go add new features for unordered list with '-' and some
bug fixes related to rendering list.

The share module add fix for permission error when scanning using memfs
and HTTP redirect for request to directory that does not end with slash.

While at it, set the minimum Go version to 1.19.